Fixing handling of big body payloads
Currently Chucker UI hangs when the HTTP request/response body is too
big. The issue is caused by having to put a bunch of text inside a
single TextView.

This commit refactors the request/response body screen to use a
RecyclerView to load single lines. In this way the UI is much smoother
and the experience is generally better.

Fixes #82