Refactor: Use regex quote for search input in `SearchHighlightUtil`
The `indicesOf` function in `SearchHighlightUtil` is refactored to use `Pattern.quote` for the search input. This ensures that any special characters in the search string are treated as literals, preventing potential `PatternSyntaxException`. In case a pattern cannot be compiled due to an invalid syntax a warning will be displayed in the logger and an empty list will be returned. Also we ensure the list is converted to immutable list to prevent unwanted modifications.