Refactor: Remove `app:popupTheme` from `activity_main_sample.xml`
The `app:popupTheme` attribute was removed from the `Toolbar` in `activity_main_sample.xml`. This change aligns the toolbar's popup theme with the application's default theme, removing a redundant setting. Based on research `content_main.xml` is used to display content, and `activity_main.xml` is used for main UI elements such as the `coordinatorLayout`. In this case, `activity_main_sample.xml` contains a `Toolbar`.