Refactor: Apply insets and update app theme (#1347)
* Refactor: Apply insets and update app theme

This commit refactors the application to handle window insets correctly, ensuring that the UI does not overlap with system bars. It also updates the application theme to remove the action bar.

-   `styles.xml`: Changed the parent theme of `AppTheme` to `Theme.MaterialComponents.DayNight.NoActionBar` to remove the default action bar.
-   `MainActivity.kt` (library): Added `applyInsets()` to handle system window insets, ensuring content is displayed properly around system bars. Added call to `applyInsets()` in `onCreate`
-   `TransactionActivity.kt`: Added `applyInsets()` to handle system window insets, ensuring content is displayed properly around system bars. Added call to `applyInsets()` in `onCreate`
-   `activity_main_sample.xml`: Added `AppBarLayout` and `MaterialToolbar` for a custom toolbar implementation.
-   `chucker_activity_transaction.xml`: Added id to `appBarLayout`
-   `MainActivity.kt` (sample): Added `applyInsets()` to handle system window insets, ensuring content is displayed properly around system bars. Added call to `applyInsets()` in `onCreate`
-   Added toolbar subtitle for sample and library `MainActivity`

* Refactor: Adjust `AppBarLayout` and UI elements in main layouts

This commit refactors the main layout files in both the library and sample modules by:

-   Removing `app:popupTheme` from `AppBarLayout` in `chucker_activity_main.xml`.
-   Moving `app:popupTheme` to `Toolbar` inside the sample's `activity_main_sample.xml`.
-   Adding missing margins to improve the UI layout consistency of `activity_main_sample.xml`.
-   Reordering views in the `activity_main_sample.xml`.
-   Adding layout horizontal margins to UI components in the `activity_main_sample.xml` such as: `description`, `interceptor_type_label`, `interceptor_type_group`, `do_http`, `do_graphql`, `launch_chucker_directly`, `export_to_file`, and `export_to_file_har`.
-   Adding a bottom margin to the `description` view in `activity_main_sample.xml`.

* Refactor: Remove `app:popupTheme` from `activity_main_sample.xml`

The `app:popupTheme` attribute was removed from the `Toolbar` in `activity_main_sample.xml`. This change aligns the toolbar's popup theme with the application's default theme, removing a redundant setting. Based on research `content_main.xml` is used to display content, and `activity_main.xml` is used for main UI elements such as the `coordinatorLayout`. In this case, `activity_main_sample.xml` contains a `Toolbar`.

* Refactor: Apply Chucker.BaseTheme to the main activity's popup theme

This commit applies the `Chucker.BaseTheme` style to the popup theme of the main activity's toolbar. This change ensures a consistent appearance for popups within the Chucker UI.

* Refactor: Update margin values in `activity_main_sample.xml`

This commit updates the horizontal margin values in the `activity_main_sample.xml` layout file to use `@dimen/doub_grid_size` instead of hardcoded `16dp`. This change makes the layout more consistent and maintainable. Additionally, it sets the bottom margin for export buttons.

* Fix: Ensure list content is visible above bottom navigation bar on Android 15

This commit addresses an issue where list and scrollable data were obscured by the bottom navigation bar on Android 15. The fix ensures that the content now scrolls to above the bottom navigation bar. The issue was resolved by updating the toolbar padding to use a defined dimension.